+++ lib/src/main/kotlin/org/didcommx/secret/SecretResolverEditable.kt
@@ -1,7 +1,5 @@
 package org.didcommx.didcomm.secret
 
-import java.util.Optional
-
 interface SecretResolverEditable : SecretResolver {
 
     fun addKey(secret: Secret)
